CROTR_Flowchart.svg added to git repository.
Panorama descriptions need some padding or margin around the text. Also move the text box in from the edge a little:

desc_background and description_Text now have Position X: 10px. desc_Text Padding: Left/Right was working on the left side only. So to get around this desc_Text was added as a child of desc_background and the Size: > Width has been changed to 90%.